Consolidate in to one folder and renaming
