+++ source/Energinet.DataHub.MeteringPoints.Application/ChangeConnectionStatus/DisconnectReconnectSettings.cs
@@ -12,7 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.MeteringPoints.Infrastructure.Integration.IntegrationEvents.Disconnect
+namespace Energinet.DataHub.MeteringPoints.Application.ChangeConnectionStatus
 {
-    public record MeteringPointDisconnectedTopic(string Name) : Topic;
+    public record DisconnectReconnectSettings
+    (
+        int NumberOfDaysEffectiveDateIsAllowedToBeforeToday = 1,
+        int NumberOfDaysEffectiveDateIsAllowedToAfterToday = 0);
 }
