+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/EDI/ChangeConnectionStatus/DisconnectReconnectMeteringPointResultHandler.cs
@@ -16,10 +16,10 @@ using System;
 using System.Linq;
 using System.Threading.Tasks;
 using Energinet.DataHub.MeteringPoints.Application;
+using Energinet.DataHub.MeteringPoints.Application.ChangeConnectionStatus;
 using Energinet.DataHub.MeteringPoints.Application.Common;
 using Energinet.DataHub.MeteringPoints.Application.Common.Commands;
 using Energinet.DataHub.MeteringPoints.Application.Connect;
-using Energinet.DataHub.MeteringPoints.Application.Disconnect;
 using Energinet.DataHub.MeteringPoints.Application.EDI;
 using Energinet.DataHub.MeteringPoints.Domain.MeteringPoints;
 using Energinet.DataHub.MeteringPoints.Domain.SeedWork;
@@ -27,7 +27,7 @@ using Energinet.DataHub.MeteringPoints.Infrastructure.BusinessRequestProcessing;
 using Energinet.DataHub.MeteringPoints.Infrastructure.EDI.AccountingPointCharacteristics;
 using Energinet.DataHub.MeteringPoints.Infrastructure.EDI.Errors;
 
-namespace Energinet.DataHub.MeteringPoints.Infrastructure.EDI.DisconnectMeteringPoint
+namespace Energinet.DataHub.MeteringPoints.Infrastructure.EDI.ChangeConnectionStatus
 {
     public class DisconnectReconnectMeteringPointResultHandler : IBusinessProcessResultHandler<DisconnectReconnectMeteringPointRequest>
     {
