+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/Energinet.DataHub.MeteringPoints.Infrastructure.csproj
@@ -63,6 +63,7 @@ limitations under the License.
 
   <ItemGroup>
     <Folder Include="EDI\Errors\Converters\Common" />
+    <Folder Include="Integration\IntegrationEvents\ChangeConnectionStatus" />
     <Folder Include="Messaging\Incoming\Protobuf\Contracts" />
   </ItemGroup>
