+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/Integration/IntegrationEvents/ChangeConnectionStatus/Disconnect/MeteringPointDisconnectedIntegrationEvent.proto
