+++ source/Energinet.DataHub.MeteringPoints.Infrastructure/Integration/IntegrationEvents/ChangeConnectionStatus/Reconnect/MeteringPointReconnectedIntegrationEvent.proto
